requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.5
django-extensions==3.2.1
django-filter==23.1
djangorestframework==3.14.0
idna==3.4
Markdown==3.4.3
psycopg2-binary==2.9.6
pydot==1.4.2
pydotplus==2.0.2
pyparsing==3.0.9
pytz==2023.3
requests==2.28.2
sqlparse==0.4.3
stripe==5.3.0
tzdata==2022.7
urllib3==1.26.15
